Palmer Penguin Dataset

A cleaned version of the Palmer Penguins dataset from the package palmerpenguins (i.e. all missing removed). This serves primarily as an example dataset for plotting examples.

palmer_penguins

Format

A tibble with 333 rows and 8 columns

species

Penguin's species

island

Penguin's island

bill_length_mm

Bill length in mm

bill_depth_mm

Bill depth in mm

flipper_length_mm

Flipper length in mm

body_mass_g

Weight in g

sex

Sex

year

Year of observation

Source

https://github.com/allisonhorst/palmerpenguins